Embed district dashboard in leader view
